test lua51

+++ openresty-nginx/src/nginx/src/os/win32/ngx_wsarecv.c
@@ -141,8 +141,8 @@ ngx_overlapped_wsarecv(ngx_connection_t *c, u_char *buf, size_t size)
     }
 
     // debug
-    output_debug_string(c, "\nngx_overlapped_wsarecv(): post event WSARecv() with buffer(0x%08xd)(%d)_bytes(%ld)ovlp(0x%08xd) on -- c(%d)fd(%d)destroyed(%d)_r(0x%08xd)w(0x%08xd)c(0x%08xd) ... w(%d)\n",
-        (uintptr_t)buf, (int)size, bytes, (uintptr_t)ovlp,
+    output_debug_string(c, "\nngx_overlapped_wsarecv(): post event WSARecv() with buffer(0x%08xd)(%d)_bytes(%d)ovlp(0x%08xd) on -- c(%d)fd(%d)destroyed(%d)_r(0x%08xd)w(0x%08xd)c(0x%08xd) ... w(%d)\n",
+        (uintptr_t)buf, (int)size, (int)bytes, (uintptr_t)ovlp,
         c->id, c->fd, c->destroyed, (uintptr_t)c->read, (uintptr_t)c->write, (uintptr_t)c, rev->write);
 
     c->buffer_ref.start = buf;
